Storing data in datatbase

+++ src/main/java/com/bl/AddressBookSpringProject/AddressBookRepository/AddressBookRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bl.AddressBookSpringProject.model.AddressBookData;
 
-public interface AddressBookRepository extends JpaRepository<AddressBookData,Long>{
+public interface AddressBookRepository extends JpaRepository<AddressBookData,Integer>{
 
 }
